requirements.txt
langchain
langchain-community
langchain-chroma
chromadb
pypdf
langgraph
sentence-transformers
numpy
pandas
fpdf
kiteconnect
openai
streamlit
plotly
scipy
yfinance
python-dotenv
